feat: add key_word.rs
